Problem: query with iavl legacy iterator is not tested
call_trace

cleanup
