+++ .gitignore
@@ -1,6 +1,7 @@
 release/
 /build
 coverage.txt
+trace.out
 __pycache__
 /dist
 /vendor
