+++ integration_tests/configs/upgrade-test-package-recent.nix
@@ -1,0 +1,26 @@
+let
+  pkgs = import ../../nix { };
+  fetchFlake =
+    repo: rev:
+    (pkgs.flake-compat {
+      src = {
+        outPath = builtins.fetchTarball "https://github.com/${repo}/archive/${rev}.tar.gz";
+        inherit rev;
+        shortRev = builtins.substring 0 7 rev;
+      };
+    }).defaultNix;
+  # release/v1.3.x
+  releasedGenesis =
+    (fetchFlake "crypto-org-chain/cronos" "e1d819c862b30f0ce978baf2addb12516568639e").default;
+  current = pkgs.callPackage ../../. { };
+in
+pkgs.linkFarm "upgrade-test-package" [
+  {
+    name = "genesis";
+    path = releasedGenesis;
+  }
+  {
+    name = "v1.4";
+    path = current;
+  }
+]
