+++ integration_tests/contracts/contracts/CheckpointOracle.sol
@@ -1,0 +1,17 @@
+// SPDX-License-Identifier: MIT
+pragma solidity ^0.8.4;
+
+contract CheckpointOracle {
+    mapping(address => bool) admins;
+    address[] adminList;
+    uint64 sectionIndex;
+    uint height;
+    bytes32 hash;
+    uint sectionSize;
+    uint processConfirms;
+    uint threshold;
+
+    function VotingThreshold() public view returns (uint) {
+        return threshold;
+    }
+}
