+++ integration_tests/utils.py
@@ -24,7 +24,8 @@ from dotenv import load_dotenv
 from eth_account import Account
 from eth_utils import abi, to_checksum_address
 from hexbytes import HexBytes
-from pystarport import ledger
+from pystarport import ledger, ports
+from pystarport.cluster import SUPERVISOR_CONFIG_FILE
 from web3._utils.contracts import abi_to_signature, find_matching_event_abi
 from web3._utils.events import get_event_data
 from web3._utils.method_formatters import receipt_formatter
@@ -63,6 +64,7 @@ TEST_CONTRACTS = {
     "TestICA": "TestICA.sol",
     "Random": "Random.sol",
     "TestRelayer": "TestRelayer.sol",
+    "CheckpointOracle": "CheckpointOracle.sol",
 }
 
 
@@ -786,3 +788,82 @@ def assert_gov_params(cli, old_param):
     expedited_param = get_expedited_params(old_param)
     for key, value in expedited_param.items():
         assert param[key] == value, param
+
+
+def init_cosmovisor(home):
+    """
+    build and setup cosmovisor directory structure in each node's home directory
+    """
+    cosmovisor = home / "cosmovisor"
+    cosmovisor.mkdir()
+    (cosmovisor / "upgrades").symlink_to("../../../upgrades")
+    (cosmovisor / "genesis").symlink_to("./upgrades/genesis")
+
+
+def post_init(path, base_port, config):
+    """
+    prepare cosmovisor for each node
+    """
+    chain_id = "cronos_777-1"
+    data = path / chain_id
+    cfg = json.loads((data / "config.json").read_text())
+    for i, _ in enumerate(cfg["validators"]):
+        home = data / f"node{i}"
+        init_cosmovisor(home)
+
+    edit_ini_sections(
+        chain_id,
+        data / SUPERVISOR_CONFIG_FILE,
+        lambda i, _: {
+            "command": f"cosmovisor run start --home %(here)s/node{i}",
+            "environment": (
+                "DAEMON_NAME=cronosd,"
+                "DAEMON_SHUTDOWN_GRACE=1m,"
+                "UNSAFE_SKIP_BACKUP=true,"
+                f"DAEMON_HOME=%(here)s/node{i}"
+            ),
+        },
+    )
+
+
+def do_upgrade(c, plan_name, target, mode=None):
+    print(f"upgrade {plan_name} height: {target}")
+    cli = c.cosmos_cli()
+    if plan_name == "v1.4.0-rc5-testnet":
+        rsp = cli.software_upgrade(
+            "community",
+            {
+                "name": plan_name,
+                "title": "upgrade test",
+                "note": "ditto",
+                "upgrade-height": target,
+                "summary": "summary",
+                "deposit": "10000basetcro",
+            },
+        )
+        assert rsp["code"] == 0, rsp["raw_log"]
+        approve_proposal(c, rsp["events"])
+    else:
+        rsp = cli.gov_propose_legacy(
+            "community",
+            "software-upgrade",
+            {
+                "name": plan_name,
+                "title": "upgrade test",
+                "description": "ditto",
+                "upgrade-height": target,
+                "deposit": "10000basetcro",
+            },
+            mode=mode,
+        )
+        assert rsp["code"] == 0, rsp["raw_log"]
+        approve_proposal(c, rsp["logs"][0]["events"])
+
+    # update cli chain binary
+    c.chain_binary = (
+        Path(c.chain_binary).parent.parent.parent / f"{plan_name}/bin/cronosd"
+    )
+    # block should pass the target height
+    wait_for_block(c.cosmos_cli(), target + 2, timeout=480)
+    wait_for_port(ports.rpc_port(c.base_port(0)))
+    return c.cosmos_cli()
